Include decision date in document title for better viewing
Right now, we record the decision date in the Git history itself
(implicitly) and the filename of the decision (explicitly), but not
actually in the *content* of the decision itself

This means that e.g.
https://ukpa-internaldocs.readthedocs.io/en/latest/decisions/ presents
you a list of decisions but without actual explicit context of when
these were (to be able to check e.g. what's new), and the only real way
to see on the published site is to look at the URL path itself, e.g.
https://ukpa-internaldocs.readthedocs.io/en/latest/decisions/2019-11-30-ukpatl-deeds/

This commit makes the dates an explicit part of the decision title,
which should both increase clarity overall, and also reduce confusion if
we ever assign the same title to a decision that was used previously
(e.g. if we make a subsequent decision pertaining to "UKPATL Deeds")